emailsend, outlook, mailgun nodes
