+++ packages/nodes-base/nodes/EmailSend/v1/EmailSendV1.node.ts
@@ -9,6 +9,7 @@ import type {
 import { NodeConnectionTypes } from 'n8n-workflow';
 import { createTransport } from 'nodemailer';
 import type SMTPTransport from 'nodemailer/lib/smtp-transport';
+import { prepareBinariesDataList } from '../../../utils/binary';
 
 const versionDescription: INodeTypeDescription = {
 	displayName: 'Send Email',
@@ -196,11 +197,7 @@ export class EmailSendV1 implements INodeType {
 
 				if (attachmentPropertyString && item.binary) {
 					const attachments = [];
-					const attachmentProperties: string[] = attachmentPropertyString
-						.split(',')
-						.map((propertyName) => {
-							return propertyName.trim();
-						});
+					const attachmentProperties = prepareBinariesDataList(attachmentPropertyString);
 
 					for (const propertyName of attachmentProperties) {
 						const binaryData = this.helpers.assertBinaryData(itemIndex, propertyName);
