+++ packages/nodes-base/nodes/Microsoft/Outlook/v2/actions/draft/create.operation.ts
@@ -4,7 +4,6 @@ import type {
 	INodeExecutionData,
 	INodeProperties,
 } from 'n8n-workflow';
-import { NodeOperationError } from 'n8n-workflow';
 
 import { updateDisplayOptions } from '@utils/utilities';
 
@@ -204,7 +203,7 @@ const displayOptions = {
 
 export const description = updateDisplayOptions(displayOptions, properties);
 
-export async function execute(this: IExecuteFunctions, index: number, items: INodeExecutionData[]) {
+export async function execute(this: IExecuteFunctions, index: number, _: INodeExecutionData[]) {
 	const additionalFields = this.getNodeParameter('additionalFields', index);
 	const subject = this.getNodeParameter('subject', index) as string;
 	const bodyContent = this.getNodeParameter('bodyContent', index, '') as string;
@@ -223,24 +222,7 @@ export async function execute(this: IExecuteFunctions, index: number, items: INo
 		body.attachments = attachments.map((attachment) => {
 			const binaryPropertyName = attachment.binaryPropertyName as string;
 
-			if (items[index].binary === undefined) {
-				throw new NodeOperationError(this.getNode(), 'No binary data exists on item!', {
-					itemIndex: index,
-				});
-			}
-
-			if (
-				items[index].binary &&
-				(items[index].binary as IDataObject)[binaryPropertyName] === undefined
-			) {
-				throw new NodeOperationError(
-					this.getNode(),
-					`No binary data property "${binaryPropertyName}" does not exists on item!`,
-					{ itemIndex: index },
-				);
-			}
-
-			const binaryData = items[index].binary[binaryPropertyName];
+			const binaryData = this.helpers.assertBinaryData(index, binaryPropertyName);
 			return {
 				'@odata.type': '#microsoft.graph.fileAttachment',
 				name: binaryData.fileName,
